requirements.txt
gunicorn
python-dotenv
flask
psycopg2
psycopg2.extras